requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<3.15
gunicorn>=20.0,<21.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=1.0,<2.0
PyJWT>=2.0,<3.0